requirements.txt
python-dotenv
requests
selenium
webdriver-manager
